third_party/blink/web_tests/platform/generic/fast/dom/Document/createElement-invalid-names-expected.txt

Test various valid name for elements according to the definition of DOM Level 2.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.createElement('0ascii') threw exception InvalidCharacterError: Failed to execute 'createElement' on 'Document': The tag name provided ('0ascii') is not a valid name..
PASS document.createElement('.Ascii') threw exception InvalidCharacterError: Failed to execute 'createElement' on 'Document': The tag name provided ('.Ascii') is not a valid name..
PASS document.createElement('-Ascii') threw exception InvalidCharacterError: Failed to execute 'createElement' on 'Document': The tag name provided ('-Ascii') is not a valid name..
PASS document.createElement(' ascii') threw exception InvalidCharacterError: Failed to execute 'createElement' on 'Document': The tag name provided (' ascii') is not a valid name..
PASS document.createElement('	ascii') threw exception InvalidCharacterError: Failed to execute 'createElement' on 'Document': The tag name provided ('	ascii') is not a valid name..
PASS document.createElement('֑sc.ii') threw exception InvalidCharacterError: Failed to execute 'createElement' on 'Document': The tag name provided ('֑sc.ii') is not a valid name..
PASS document.createElement('⃣scii') threw exception InvalidCharacterError: Failed to execute 'createElement' on 'Document': The tag name provided ('⃣scii') is not a valid name..
PASS document.createElement('िascii') threw exception InvalidCharacterError: Failed to execute 'createElement' on 'Document': The tag name provided ('िascii') is not a valid name..
PASS document.createElement('ʳascii') threw exception InvalidCharacterError: Failed to execute 'createElement' on 'Document': The tag name provided ('ʳascii') is not a valid name..
PASS document.createElement('asc i') threw exception InvalidCharacterError: Failed to execute 'createElement' on 'Document': The tag name provided ('asc i') is not a valid name..
PASS document.createElement('asc	i') threw exception InvalidCharacterError: Failed to execute 'createElement' on 'Document': The tag name provided ('asc	i') is not a valid name..
PASS document.createElement('asciiⅦ') threw exception InvalidCharacterError: Failed to execute 'createElement' on 'Document': The tag name provided ('asciiⅦ') is not a valid name..
PASS successfullyParsed is true

TEST COMPLETE
